// src/types/lol.ts
export type Lane = 'Top' | 'Jungle' | 'Mid' | 'Bot' | 'Support';

export type Team = 'Blue' | 'Red';

export type RankedTier =
	| 'CHALLENGER'
	| 'GRANDMASTER'
	| 'MASTER'
	| 'DIAMOND'
	| 'EMERALD'
	| 'PLATINUM'
	| 'GOLD'
	| 'SILVER'
	| 'BRONZE'
	| 'IRON';

export type RankedDivision = 'I' | 'II' | 'III' | 'IV';